requirements.txt
streamlit==1.39.0
Pillow==10.3.0
openai==0.27.0
PyPDF2==3.0.1
pandas==2.2.3
matplotlib==3.8.4
faiss-cpu==1.9.0
seaborn==0.13.2
pygame==2.6.1
python-dotenv==1.0.1
tiktoken==0.7.0
python-docx==1.1.2